API

WeightedOnlineStats.WeightedAdaptiveHist

WeightedOnlineStats.WeightedCovMatrix

WeightedOnlineStats.WeightedHist1D

WeightedOnlineStats.WeightedMean

WeightedOnlineStats.WeightedSum

WeightedOnlineStats.WeightedVariance

WeightedOnlineStats.pca

WeightedOnlineStats.WeightedAdaptiveHist — Type

Weighted Histogram

Calculate a histogram of weighted data.

Example

# A weighted histogram with 4 bins:
o = fit!(WeightedAdaptiveHist(4), rand(1000), rand(1000))

mean(o)
var(o)
std(o)
median(o)
quantile(o, [0, 0.25, 0.5, 0.25, 1.0])
extrema(o)

WeightedOnlineStats.WeightedCovMatrix — Type

WeightedCovMatrix(T = Float64)

Weighted covariance matrix, tracked as a matrix of type T.

After a call to cov the covariance matrix is stored in o.C.

Example:

o = fit!(WeightedCovMatrix(), rand(100, 4) |> eachrow, rand(100))
o = fit!(WeightedCovMatrix(), rand(4, 100) |> eachcol, rand(100))
sum(o)
mean(o)
var(o)
std(o)
cov(o)
cor(o)

WeightedOnlineStats.WeightedHist1D — Type

WeightedHist(edges; left = true, closed = true)

Create a histogram with bin partition defined by edges.

If left, the bins will be left-closed.

If closed, the bin on the end will be closed.

E.g. for a two bin histogram $[a, b), [b, c)$ vs. $[a, b), [b, c]$

If edges is a tuple instead of an array, a multidimensional histogram will be generated that behaves like a WeightedOnlineStat{VectorOb}.

Examples

o = fit!(WeightedHist(-5:.1:5), randn(10^6))

# approximate statistics
using Statistics
mean(o)
var(o)
std(o)
quantile(o)
median(o)
extrema(o)
area(o)
pdf(o)

2d Histogram

hist2d = fit!(WeightedHist((-5:1:5, -5:1:5) ), randn(10000,2), rand(10000))
value(hist2d).y

WeightedOnlineStats.WeightedMean — Type

WeightedMean(T = Float64)

Simple weighted mean, tracked as type T.

Example:

o = fit!(WeightedMean(), rand(100), rand(100))
sum(o)
mean(o)

WeightedOnlineStats.WeightedSum — Type

WeightedSum(T = Float64)

Simple weighted sum, tracked as type T.

Example:

o = fit!(WeightedSum(), rand(100), rand(100))
sum(o)

WeightedOnlineStats.WeightedVariance — Type

WeightedVariance(T = Float64)

Simple weighted variance, tracked as type T.

Example:

o = fit!(WeightedVariance(), rand(100), rand(100))
sum(o)
mean(o)
var(o)
std(o)

WeightedOnlineStats.pca — Method

pca(x::WeightedCovMatrix)::Tuple{StatsBase.ZScoreTransform, MultivariateStats.PCA}

Constructs a StatsBase.ZScoreTransform, and a MultivariateStats.PCA object from a WeightedCovMatrix.

Parameters

cov_pca::Bool = false: Do a covariance or correlation pca.

maxoutdim::Int = size(x, 1): Limits the maximum number of output dimensions.

pratio::AbstractFloat = 1.0: Limits the number of output dimensions to a ratio of explained variance.

correct::Bool = false: Correct the cov(x)/cor(x)/std(x), see the documentation of weights in StatsBase.jl for details.

weight_type::Symbol: The type of weight for the correction, see the documentation of weights in StatsBase.jl for details. Ignored if correct is false.

Example

c = fit!(WeightedCovMatrix(), rand(4, 100), rand(100))
t, p = pca(c)